w2s update (only vertically accurate)

+++ ToadClient/Toad/MC/Utils/utils.h
@@ -96,7 +96,7 @@ namespace toadll
     float wrap_to_180(float value);
 
     //bool WorldToScreen(const vec3& worldpos, vec2& screen, GLfloat modelView[15], GLfloat projection[15], GLint viewPort[3]);
-    bool WorldToScreen(const vec3& source, const vec3& target, const vec2& viewAngles, int fov, vec2& screenpos);
+    bool WorldToScreen(const vec3& source, const vec3& target, const vec2& viewAngles, float fov, vec2& screenpos);
 
     void loop_through_class(const jclass klass);
